Root Finding/requirements.txt
sympy
multipledispatch
pandas
Pillow
ttkthemes